requirements.txt
boto3==1.34.131
google-api-python-client==2.128.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
paramiko==3.4.0
PyYAML==6.0.2
requests==2.32.3
